+++ core/src/main/java/de/fraunhofer/ids/messaging/core/config/ConfigContainer.java
@@ -102,14 +102,14 @@ public class ConfigContainer {
         } catch (KeyStoreManagerInitializationException e) {
             if (log.isErrorEnabled()) {
                 log.error("Configuration could not be updated!"
-                    + " Keeping old configuration! {}", e.getMessage());
+                    + " Keeping old configuration! [exception=({})]", e.getMessage());
             }
 
             throw new ConfigUpdateException(e.getMessage(), e.getCause());
         } catch (NoSuchAlgorithmException | KeyManagementException e) {
             if (log.isErrorEnabled()) {
                 log.error("New Key- or Truststore could not be initialized!"
-                    + " Keeping old configuration! {}", e.getMessage());
+                    + " Keeping old configuration! [exception=({})]", e.getMessage());
             }
             throw new ConfigUpdateException(e.getMessage(), e.getCause());
         }