+++ core/src/main/java/de/fraunhofer/ids/messaging/core/daps/customvalidation/DatValidationRule.java
@@ -22,6 +22,7 @@ import io.jsonwebtoken.Claims;
 public interface DatValidationRule {
     /**
      * Adds the possibility to add custom validation rules.
+     *
      * @param toVerify The Claims to verify.
      * @return ValidationRuleResult if successful or not and message.
      * @throws ValidationRuleException If exception thrown by custom check.