+++ messaging/src/main/java/de/fraunhofer/ids/messaging/protocol/multipart/mapping/ContractAgreementMAP.java
@@ -27,6 +27,7 @@ public class ContractAgreementMAP
 
     /**
      * Constructor for ContractAgreementMAP.
+     *
      * @param contractAgreementMessage The message.
      * @param payload The payload.
      */