+++ messaging/src/main/java/de/fraunhofer/ids/messaging/protocol/multipart/mapping/ContractRequestMAP.java
@@ -28,6 +28,7 @@ public class ContractRequestMAP
 
     /**
      * Constructor for ContractRequestMAP.
+     *
      * @param contractRequestMessage The message.
      * @param payload The payload.
      */