+++ messaging/src/main/java/de/fraunhofer/ids/messaging/protocol/multipart/mapping/ParticipantRequestMAP.java
@@ -31,6 +31,7 @@ public class ParticipantRequestMAP implements MessageAndPayload<ParticipantReque
 
     /**
      * Constructor for the ParticipantRequestMAP.
+     *
      * @param message The ParticipantRequestMessage.
      */
     public ParticipantRequestMAP(final ParticipantRequestMessage message) {