+++ messaging/src/main/java/de/fraunhofer/ids/messaging/util/RequestUtils.java
@@ -48,7 +48,7 @@ public final class RequestUtils {
             buffer.close();
         } catch (IOException | NullPointerException e) {
             if (log.isErrorEnabled()) {
-                log.error("{}: printing failed.", e.getClass().toString());
+                log.error("Printing failed! [error=({})]", e.getClass().toString());
             }
         }