+++ CHANGELOG.md
@@ -9,6 +9,23 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Version [4.3.0] 2021-08-31
+
+### Minor Change: Incoming message infomodel compatibility check
+- The check whether an incoming message is compatible with its ModelVersion to the inbound ModelVersions of the Connector can be switched on or off via application.properties (infomodel.compatibility.validation=true/false). The default value if not set is true (switched on).
+
+### Patch Change: Infomodel Maintenance
+- Used Dependency Version: 4.2.1 (released 2021-08-30)
+- Used Artifacts: java, infomodel-serializer, interaction
+
+### Patch Change: Enhancements
+- If no private key can be found for a given alias within a keystore, a KeyStoreException is now thrown and an error is logged, preventing an otherwise possible NullpointerException (KeystoreManager getPrivateKeyFromKeyStore).
+- The payload of incoming messages is now first validated for valid JSON and whether the securityProfile attribute is present, if not, this check is skipped instead of issuing an error message in the logs (IdsHttpService checkDatFromResponse).
+- For log error messages and others important logs with dynamic content, the e.g. exception reason is now highlighted to distinguish it from the rest of the log message. Format: [exception=(...)].
+
+### Patch Change: Miscellaneous
+- Other minor enhancements to existing Javadoc and log messages.
+
 ## Version [4.2.2] 2021-08-26
 
 ### Patch Change: Miscellaneous